policy: use output.virtual-source nodes if any, and hide them
This hack improves the policy to use output.virtual-source node by default if
they are available. Those nodes are then linked to the filer-chain nodes if they
are available, otherwise the will be linked to the echo-cancel nodes. If
echo-cancel nodes are not available, the real audio devices will be used.

One the other hand, similar to the previous hiding of echo-cancel nodes hack,
this patch also hides the output.virtual-source node so that some applications
don't get confused when getting the real default device node.

This allows applications to always use the same source, regardeless of whether
echo-cancel/filter-chain nodes are used or not.